backend-mouse-kerramientas/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=1.8.2
pydantic-settings>=2.0.0
alembic>=1.7.1
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
bcrypt>=3.2.0
email-validator>=1.1.3
python-dotenv>=0.19.0
pytest>=6.2.5
pytest-asyncio>=0.18.0
httpx>=0.23.0